ReadMe file
I believe this is the README file changes and it is placing it in the
workspace. You should re-sync to get this. Also, go to GitHub on the web
if you want to just download the readme file. It contains important
information about pulling and pushing update to GitHub. Everybody should
go here first if you have any problems and/or forgot what to do