Added HttpURLConnection Code
This commit will sync you with all of the HttpURLConnection code i
implemented into the project. It contains many...MANY error fixes. It
does not fully work yet as no new users are added to the database,
however, I feel confident it is very close. Sync to this if you like or
wait until the final fix that makes it work. Coming Soon to theaters
near you (hopefully.)