Working Project - Working Database Code
**IMPORTANT** - Gentlemen, this is the working project as of 11/6, that
will successfully connect to the database and add users. Please sync to
this version and we can review all of the changes I made the next time
we meet. I plan on making a list of things to do so that we can pick and
delegate work and have a deadline that we should all complete our tasks
by.