Task 3 - Completion
Added code to display a message depending on whether or not an account
was created successfully. If the server response is good, then the user
is directed back to the login page and a success message is displayed.
If the server response is bad, then the user stays on the account
creation page and an error message is displayed, telling the user what
to do next.