Functionality for global variables, list of user files in FilesActivity
I added a class called Globals that allows us to declare and use global
variables. Description on how to use it is in the class commented out.
Also on FilesActivity a list of file names now comes up or a
notification saying there are no files. To test the list log in with the
username "kartoone" and the password "testpassword" as that is the only
user that currently has any new files.