Shows file contents.
On the files activity it now shows the file contents of the selected
file to the right of the list of files. I tried to adjust the size of
the text field to make room for a label that would say "file contents"
or something like that but it literally would make it as small as
possible no matter what I set it to. Then I tried moving the create
files button over to the left to make room for the label but for some
reason instead of just moving the button it would only let me move
everything on the entire page. So I gave up on that and left it how it
is. Stupid android studio.