Fully Functional AddUserToFileActivity. Also added support for editable/read-only permissions
This update commits the fully functional AddUserToFileActivity page. It
now makes the necessary calls and adds the specified user to the
specified file. The actual database call to addusertofile.php requires a
UserId and FileId, but i put in the extra work to actually allow the
user to enter in a username and filename instead. Then behind the scenes
i make some extra calls and convert those names to the id's in the
database for the actual database call.

You will also notice the permissions: "read-only" or "can edit". Because
i had to add this option in the new activity for users to select, i went
ahead and added support for this in the FilesActivity list. It now
disables the fileContentBox if the file happens to be a shared file and
has a permission of 0 (read-only). It will enable the fileContentBox
which allows editing for shared files that have a permission of 1. This
wraps up everything i needed to do for my task.