Added ability  to create a new file
When you click the create file button in the files activity it takes you
to a new activity where you can enter the name for a file and enter some
contents. Then when you click the create file button on that activity it
adds the file to the database and takes you back to the files activity
where you will see the created file in the list of files. We will have
to do the colors on the create file activity though.