Further Group Support Improvements...See notes
This commit has a lot of new improvements for group support. First,
there is now a distinction between owner and member when displaying a
user's groups. Second, the group owner is identified in the member list.
Third, A blank group name can no longer be entered. Fourth, a member of
a group cannot delete  a group that they do not own (that is reserved
for group owners only, currently).